backend/requirements.txt
amqp==5.0.6
appnope==0.1.2
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
billiard==3.6.4.0
boto3==1.18.28
botocore==1.21.28
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.0.9
Django==3.2.6
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-guardian==2.4.0
django-nine==0.2.4
django-ratelimit==3.0.1
django-redis==5.0.0
django-rest-framework-tricks==0.2.11
django-silk @ git+https://github.com/jazzband/django-silk@76a9b1e5dd173fce4216aad33a5dc7f9954c5f94
django-stdimage==5.3.0
django-storages==1.11.1
djangorestframework==3.12.4
drf-extensions==0.7.1
drf-nested-routers==0.93.3
Faker==8.11.0
flower==1.0.0
gevent==21.8.0
gprof2dot==2021.2.21
greenlet==1.1.1
gunicorn==20.1.0
hiredis==2.0.0
humanize==3.11.0
idna==3.2
ipython==7.26.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
python-dateutil==2.8.2
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
sendgrid==6.8.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
